% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin_chrom}
\alias{bin_chrom}
\title{Bin an existing InteractionSet}
\usage{
bin_chrom(Iset, binSize)
}
\arguments{
\item{Iset}{An InteractionSet object.}

\item{binSize}{The desired bin size.}
}
\value{
The binned InteractionSet, with the bin size stored in metadata(myInteractionSet)$binSize
}
\description{
For any InteractionSet object, bin the regions and the interactions to make them more feasible to visualize.
}

