% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr2}
\alias{chr2}
\title{Hi-C interaction data for chromosome 2 of mouse stem cells}
\format{An InteractionSet object}
\source{
Pekowska, A., Klaus, B., Klein, F. A., Anders, S., Oles, M., Steinmetz, L. M., … Huber, W. (2014).
Neural lineage induction reveals multi-scale dynamics of 3D chromatin organization. bioRxiv.
Cold Spring Harbor Labs Journals. Retrieved from http://biorxiv.org/content/early/2014/04/11/004168.abstract
}
\usage{
chr2
}
\description{
Interactions of the chromosome 2 with all other chromosomes
}
\keyword{datasets}

