% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{easy_Iset_to_LFM}
\alias{easy_Iset_to_LFM}
\title{Convert an InteractionSet into a sparse, upper triangle ligation frequency matrix.}
\usage{
easy_Iset_to_LFM(Iset, chr = NULL, from = NULL, to = NULL)
}
\arguments{
\item{Iset}{An interactionSet object, with the bin size stored as metadata under the name 'binSize'.}

\item{chr}{Optional. The chromosome name to be plotted with or without 'chr', depending on your interactionSet object.}

\item{from}{Optional. Start of the region you want to select. If NULL, the start of the chromosome will be 'from'.}

\item{to}{Optional. End of the region you want to select. If NULL, the end of the chromosome will be 'to'.}
}
\value{
A sparse, upper triangle ligation frequency matrix.
}
\description{
Wrapper for Iset_to_LFM (no subsetting) and Iset_region_to_LFM (subsetting by chromosome and, if required, by region).
}
\examples{
LFM <- easy_Iset_to_LFM(Iset)
chr2_1_300000 <- easy_Iset_to_LFM(Iset, chr = 2, from = 1, to = 300000)

}

