% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_subset.R
\name{import_chrom}
\alias{import_chrom}
\title{Read preprocessed genomic interaction files into an InteractionSet object.}
\usage{
import_chrom(bed, raflist = NULL, workDir = getwd(), binned = FALSE)
}
\arguments{
\item{bed}{File name of a bed file without header, with columns for chromosome (Ensembl format, no "chr"), start, end, and
restriction fragment ID of each restriction fragment.}

\item{raflist}{A list of file names containing the interactions. The files are in raf format, meaning they contain two
columns (no header) with the interacting restriction fragment IDs. If raflist=NULL, all raf files in the directory will
be read.}

\item{workDir}{The directory where all input files are located. Defaults to the current working directory.}

\item{binned}{Boolean, whether the matrix is already binned. Defaults to FALSE.}
}
\value{
An InteractionSet object containing all interactions from all input raf files, and the bin size (if any) in the metadata.
}
\description{
Read preprocessed genomic interaction files into an InteractionSet object.
}
\examples{
# You can import several specific raf files from the same directory
# (the working directory if no workDir is specified).
# myInteractions <- import_chrom(bed = "annotation.bed",
#                                raflist = list("1.raf", "2.raf"))

# Or just choose all raf files within a specific directory
myInteractions <- import_chrom(bed = "RFanno_HindIII.bed", workDir = system.file("extdata", package = "chromium"), binned = F)

}

