% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{IPF_alg}
\alias{IPF_alg}
\title{Iterative proportional fitting algorithm}
\usage{
IPF_alg(symLFM, numberOfIterations)
}
\arguments{
\item{symLFM}{A symmetric sparse ligation frequency matrix (LFM) with binned or unbinned data.}

\item{numberOfIterations}{The number of iterations for normalization. Usually, 10 should be more than enough.}
}
\value{
A list containing the normalized, symmetric matrix \code{normLFM}, the dataframes with the row sums
\code{rowBiases} and the column sums \code{colBiases} of every iteration, and a vector with the loss function \code{lf}
for every iteration \code{lossFunction}.
}
\description{
Iterative proportional fitting (IPF) is used to normalize a sparse matrix.
}
\examples{
M <- matrix(c(3,0,1,0,7,9,1,9,7),3,3)
M <- Matrix::Matrix(M, sparse = TRUE)
ipf <- chromium:::IPF_alg(M,10)

}

