% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{Iset_region_to_LFM}
\alias{Iset_region_to_LFM}
\title{Convert an InteractionSet region to a ligation frequency matrix.}
\usage{
Iset_region_to_LFM(Iset, chr, from = NULL, to = NULL)
}
\arguments{
\item{Iset}{An interactionSet object, with the bin size stored as metadata under the name 'binSize'.}

\item{chr}{The chromosome name to be plotted with or without 'chr', depending on your interactionSet object.}

\item{from}{Start of the region you want to select. If NULL, the start of the chromosome will be 'from'.}

\item{to}{End of the region you want to select. If NULL, the end of the chromosome will be 'to'.}
}
\description{
Pick a genomic region from an InteractionSet object and convert it into a sparse triangular matrix.
}
\examples{
matrix <- Iset_region_to_matrix(Iset, chr = 11, from = 30000000, to = 30100000)

}

