% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{loss_function}
\alias{loss_function}
\title{Loss function}
\usage{
loss_function(rowsums, colsums)
}
\arguments{
\item{rowsums}{The current row sums of the matrix}

\item{colsums}{The current column sums of the matrix}
}
\value{
A decimal number giving the loss function of the current row and column sums.
}
\description{
Loss function of the iterative proportional fitting algorithm
}
\examples{
M <- matrix(c(3,0,1,0,7,9,1,9,7),3,3)
M <- Matrix::Matrix(M, sparse = TRUE)
lf <- loss_function(Matrix::rowSums(M), Matrix::colSums(M))
}

