% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize}
\alias{normalize}
\title{Normalize an InteractionSet with iterative proportional fitting.}
\usage{
normalize(Iset, numberOfIterations = 40L, plotLoss = FALSE)
}
\arguments{
\item{Iset}{An InteractionSet with binned or unbinned data. If the data is binned, metadata(Iset)$binSize must contain
the bin size.}

\item{numberOfIterations}{The number of iterations for normalization. Usually, 20 should be more than enough.}

\item{plotLoss}{Optional boolean parameter, to determine if the loss function should be plotted. Defaults to FALSE.}
}
\value{
A list containing the normalized InteractionSet, i.e. assay(normIset) will contain the normalized signal. In the
metadata, the number of interchromosomal contacts and intrachromosomal contacts is added, as well as the contact distances,
i.e. the distances between the genomic positions where two interacting bins or restriction fragments start. In addition,
the loss function is given for every iteration of the normalization. If plotLoss = T, the metadata will also contain a
plotted version of the loss function, which can be displayed with metadata(normIset)$lossPlot.
}
\description{
A binned or unbinned InteractionSet is normalized using iterative proportional fitting (IPF), an implementation of the
iterative matrix balancing algorithm by Pukelsheim and Simeone (2009).
}
\examples{
norm_Iset <- normalize(raw_Iset, numberOfIterations = 40L, plotLoss = FALSE)

}

