% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_LFM_iteratively}
\alias{normalize_LFM_iteratively}
\title{Normalize a ligation frequency matrix with iterative proportional fitting}
\usage{
normalize_LFM_iteratively(LFM, numberOfIterations, plotLoss = FALSE)
}
\arguments{
\item{LFM}{A sparse ligation frequency matrix (LFM) with binned or unbinned data.}

\item{numberOfIterations}{The number of iterations for normalization. Usually, 20 should be more than enough.}

\item{plotLoss}{Optional boolean parameter, to determine if the loss function should be plotted. Defaults to FALSE.}
}
\value{
A list containing the normalized, triangular matrix \code{finalLFM}, the dataframes with the row sums
\code{IPFrowsums} and the column sums \code{IPFcolsums} of every iteration, and a vector with the loss function
for every iteration \code{lossFunction}.
}
\description{
A binned or unbinned ligation frequency matrix (LFM) is normalized using iterative proportional fitting (IPF).
}
\examples{
M <- matrix(c(3,0,0,0,7,0,1,9,7),3,3)
M <- Matrix(M, sparse = TRUE)
normalizedLFM <- normalize_LFM_iteratively(M, 10)

}

