% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{pairID_to_RFpair}
\alias{pairID_to_RFpair}
\title{Decode the pair ID into the bin pairs or restriction fragment pairs.}
\usage{
pairID_to_RFpair(pairID, RFanno)
}
\arguments{
\item{pairID}{A vector with pair IDs that each encode a pair of restriction fragments or bins.}

\item{RFanno}{A dataframe giving the genomic annotation of the restriction fragments or bins. Columns are: chromosome, start of
restriction fragment or bin, end of restriction fragment or bin, restriction fragment or bin ID. The table is sorted by chromosomal
position and restriction fragments or bins are numbered along their genomic position.}
}
\value{
A dataframe with two columns containing restriction fragment pairs or bin pairs.
}
\description{
Using the genomic annotation, the unique pair ID is decoded into the two restriction fragment IDs or bin IDs.
}
\examples{
# pair IDs are given here
code <- c(16,47,78,109,140,171,202,233,264,295,326,357,388,419,450)

# genomic annotation of the restriction fragment pairs
chr <- c(rep.int(1,10), rep.int(2,10), rep.int(3,10))
start <- rep.int(seq(1, 30, by = 3),3)
end <- start + 2
RF_id <- c(1:30)
myanno <- data.frame(chr, start, end, RF_id)

# generate the decoded pairs
mypairs <- pairID_to_RFpair(code, myanno)

> mypairs
      [,1] [,2]
[1,]    1   16
[2,]    2   17
[3,]    3   18
 ..     ..  ..

}

