% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assessment.R
\name{plot_qa}
\alias{plot_qa}
\title{Output a quality assessment report.}
\usage{
plot_qa(qa, outDir = getwd(), reportName = "Quality_report.pdf")
}
\arguments{
\item{qa}{The list from the quality_assessment() function.}

\item{outDir}{The directory where the quality report should be created.}

\item{reportName}{File name of the output PDF file}
}
\description{
Outputs a PDF with number of ligation products (treated as interactions), number of inter- and intrachromosomal interactions,
and plots of distances between interactions and the loss function of normalization (if data are normalized).
}
\examples{
plot_qa(qa, reportName = "My_qality_report.pdf")

}

