% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assessment.R
\name{quality_assessment}
\alias{quality_assessment}
\title{Quality assessment of whole genome Hi-C data.}
\usage{
quality_assessment(RFpairs, RFanno, Iset = NULL)
}
\arguments{
\item{RFpairs}{A dataframe with two columns containing restriction fragment pairs. The restriction fragment with the lower ID
is always in the left column, and all restriction fragments are sorted in ascending order first by the left column and then by
the right colunn.}

\item{RFanno}{A dataframe giving the genomic annotation of the restriction fragments. Columns are: chromosome, start of
restriction fragment, end of restriction fragment, restriction fragment ID. The table is sorted by chromosomal position
and restriction fragments are numbered along their genomic position.}

\item{Iset}{An InteractionSet object with lossFunction from normalization in its metadata.}
}
\value{
A list with the quality assessment values interactionCount, interchromosomal, intrachromosomal, contactDistances,
and if it exist, the lossFunction values of each iteration of the normalization.
}
\description{
Calculate number of ligation products (treated as interactions), number of inter- and intrachromosomal interactions,
and distances between interactions.
}
\details{
here go the details
}
\examples{
example
}

