% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_subset.R
\name{read_bed_raf}
\alias{read_bed_raf}
\title{Read preprocessed genomic interaction files to produce RFanno and RFpairs.}
\usage{
read_bed_raf(bed, raflist = NULL, workDir = getwd())
}
\arguments{
\item{bed}{File name of a bed file without header, with columns for chromosome (Ensembl format, no "chr"), start, end, and
restriction fragment ID of each restriction fragment.}

\item{raflist}{A list of file names containing the interactions. The files are in raf format, meaning they contain two
columns (no header) with the interacting restriction fragment IDs. If raflist=NULL, all raf files in the directory will be
read.}

\item{workDir}{The directory where all input files are located. Defaults to the current working directory.}
}
\value{
A list consisting of an RFanno dataframe containing the restriction fragment annotation, and an RFpairs dataframe
containing the interacting restriction fragment pairs.
}
\description{
Read preprocessed genomic interaction files to produce RFanno and RFpairs.
}
\examples{
RFanno <- read_bed_raf(bed = "annotation.bed", raflist = list("1.raf", "2.raf"))[[1]]
RFpairs <- read_bed_raf(bed = "annotation.bed", raflist = list("1.raf", "2.raf"))[[2]]

}

