% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{RFpairs_to_Iset}
\alias{RFpairs_to_Iset}
\title{Convert restriction fragment annotation and restriction fragment pair objects to an InteractionSet object.}
\usage{
RFpairs_to_Iset(RFanno, RFpairs, binned = F)
}
\arguments{
\item{RFanno}{A dataframe giving the genomic annotation of the restriction fragments. Columns are: chromosome, start of
restriction fragment, end of restriction fragment, restriction fragment ID. The table is sorted by chromosomal position
and restriction fragments are numbered along their genomic position.}

\item{RFpairs}{A dataframe with two columns containing restriction fragment pairs. The restriction fragment with the
lower ID is always in the left column, and all restriction fragments are sorted in ascending order first by the left
column and then by the right colunn.}

\item{binned}{Boolean, whether the matrix is already binned. Defaults to FALSE.}
}
\value{
An InteractionSet object.
}
\description{
Convert restriction fragment annotation and restriction fragment pair objects to an InteractionSet object.
}
\examples{
# Iset <- RFpairs_to_Iset(RFanno, RFpairs)

}

