% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{rotated_image}
\alias{rotated_image}
\title{Plot a rotated and, if required, smoothed interaction matrix.}
\usage{
rotated_image(img, scaleCol, smooth, smoothing, filterSize, sigma, binSize)
}
\arguments{
\item{img}{The symmetric, non-sparse interaction matrix (for feasibility, choose only a limited chromosomal region).}

\item{scaleCol}{The color scale threshold. All values above the threshold will ne shown in the highest intensity color.}

\item{smooth}{Which smoothing type should be applied? Options are: "none", "box", and "gaussian".}

\item{smoothing}{Smoothing parameter. The percentage of signal to be taken from the surrounding pixels. Smoothing is optional.}

\item{filterSize}{Smoothing parameter. The number of pixels surrounding the central pixel. (Size of filter mask.) Smoothing is optional.}

\item{sigma}{In the case of Gaussian smoothing, how large should the sigma of the distribution be?}

\item{binSize}{Size of bins into which the interaction matrix is binned.}
}
\description{
A matrix with the interactions is rotated and cut along the diagonal to form a triangle plot of the interactome. Smoothing
is performed utilizing the filter2 convolution filter from the EBImage package.
}
\examples{
myTriangle <- rotated_image(img, scaleCol = 0.02, smoothing = 80, filterSize = 3, binSize = 10000)

}

