% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_subset.R
\name{subset_chrom}
\alias{subset_chrom}
\title{Choose a chromosome and a region within which InteractionSet interactions and regions should be retained.}
\usage{
subset_chrom(Iset, chr, from = NULL, to = NULL)
}
\arguments{
\item{Iset}{The InteractionSet object to be subsetted.}

\item{chr}{The chromosome for which interactions and regions should be retained.}

\item{from}{Optional. The starting point of the interactions and regions to be retained. If NULL, the chromosome
beginning is selected.}

\item{to}{Optional. The ending point of the interactions and regions to be retained. If NULL, the chromosome end
is selected.}
}
\value{
An InteractionSet containing only those regions and those interactions specified in the arguments.
}
\description{
Only those interactions and regions of an InteractionSet which lie within the specified range will be retained.
This is useful to reduce the object size of the InteractionSet.
}
\examples{
# chr2_Iset <- subset_chrom(Iset, chr = 2)

}

