% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_subset.R
\name{subset_interactions}
\alias{subset_interactions}
\title{Choose a chromosome and a range within which InteractionSet interactions should be retained.}
\usage{
subset_interactions(Iset, chr, from = NULL, to = NULL)
}
\arguments{
\item{Iset}{The InteractionSet object to be subsetted.}

\item{chr}{The chromosome for which interactions should be retained.}

\item{from}{Optional. The starting point of the interactions to be retained. If NULL, the chromosome beginning is selected.}
}
\value{
to Optional. The ending point of the interactions to be retained. If NULL, the chromosome end is selected.
}
\description{
The interactions of the InteractionSet will be subsetted, but all regions (i.e. the annotation) will be retained.
This is only useful if you want to add interactions in the other regions later. Otherwise, especially for direct
conversion to matrices, it is more useful to subset the whole InteractionSet, including the regions.
}
\examples{
chr2_Iset <- subset_interactions(Iset, chr=2)

}

