% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{sym_to_triangle}
\alias{sym_to_triangle}
\title{Convert a sparse symmetric matrix to a sparse triangle matrix.}
\usage{
sym_to_triangle(mat)
}
\arguments{
\item{mat}{A symmetric sparse matrix.}
}
\value{
A triangular sparse matrix, containing only the upper triangle.
}
\description{
Convert a sparse symmetric matrix to a sparse triangle matrix.
}
\examples{
triangleMatrix <- sym_to_triangle(symmMatrix)

}

