% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_chrom}
\alias{visualize_chrom}
\title{Triangle visualization of Hi-C data with tracks for chipseq, genome axis, gene models, and custom data.}
\usage{
visualize_chrom(Iset, chr, from = NULL, to = NULL, gen = NULL,
  geneModels = NULL, printTranscripts = TRUE, outDir = getwd(),
  name = "triangle_visualization.pdf", scaleCol = 0.02, chipseqs = NULL,
  chipScale = NULL, colmapChipseqs = NULL, customAnno = NULL,
  smooth = "none", smoothing = NULL, filterSize = NULL, sigma = NULL)
}
\arguments{
\item{Iset}{An interactionSet object, with the bin size stored as metadata under the name "binSize"}

\item{chr}{The chromosome name to be plotted with or without "chr", depending on your interactionSet object.}

\item{from}{Start of the region you want to plot.}

\item{to}{End of the region you want to plot.}

\item{gen}{The genome identifier, used primarily to determine which gene model will be used.
Give the organism name (e.g. "hsapiens", "mmusculus") if you want to use the most recent assembly from biomart (internet
connection required). Give one of "mm9", "mm10", "hg19", or "hg38" to use a gene model bundled with the package.
Leave out if you want to use your own gene model.}

\item{geneModels}{Only input a geneModel GRanges if you want to use your own custom geneModel.}

\item{printTranscripts}{Boolean. Produce gene model track? Defaults to TRUE.}

\item{outDir}{Give the directory where the PDF plot will be saved. Defaults to the working directory.}

\item{name}{A string. Filename of the output PDF.}

\item{scaleCol}{The color scale threshold. All values above the threshold will ne shown in the highest intensity color.}

\item{chipseqs}{A named list of GRanges objects with ChIP-seq scores in the metadata column "score". Optional.}

\item{chipScale}{A named list corresponding to chipseqs with a vector containing the highest and lowest score for each list
element. Defaults to taking the non-zero 99\% quantile (rounded to the nearest ten) as the maximum and -0.5 as the minimum value.}

\item{colmapChipseqs}{A named character vector corresponding to chipseqs with a color name for each chipseq. Defaults to a
vector of twelve colors. For a larger number of chipseq tracks, please define colmapChipseqs.}

\item{customAnno}{A custom annotation, given as a list of BED dataframes. Optional.}

\item{smooth}{Which smoothing type should be applied? Options are: "none", "box", and "gaussian".}

\item{smoothing}{Smoothing parameter. The percentage of signal to be taken from the surrounding pixels. Smoothing is optional.}

\item{filterSize}{Smoothing parameter. The dimension of filter mask, e.g. filterSize = 3 means 3x3 filter. Smoothing is optional.}

\item{sigma}{In the case of Gaussian smoothing, how large should the sigma of the distribution be?}
}
\value{
Output will be in PDF format.
}
\description{
An interaction matrix is plotted, rotated, and if required, smoothed. Using Gviz, tracks can be added as required.
}
\examples{
visualize_chrom(Iset, chr = "11", from = 30000000, to = 30100000,
          gen = "mm9", customAnno = cpgIslands, smoothing = 80, filterSize = 3)

}

