% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{Iset_to_LFM}
\alias{Iset_to_LFM}
\title{Convert an InteractionSet object into a ligation frequency matrix.}
\usage{
Iset_to_LFM(Iset)
}
\arguments{
\item{Iset}{An InteractionSet object}
}
\value{
A list consisting of: the sparse ligation frequency matrix with the genomic annotation as rownames and colnames,
the restriction fragment annotation (or bin annotation), and the restriction fragment pairs (or bin pairs).
}
\description{
The data can be binned or not binned, normalized or not normalized.
}
\examples{
LFM=Iset_to_LFM(Iset)[[1]]

}

